--- src/js/main.js
@@ -1,1 +1,0 @@
-require('./dom-scripts.js');
